Finally! Graphics!
Finally! We got graphics!

The example app is able to create sprites which are rendered by GLHF every frame. Still has some issues at the moment though. One immediate example is that the sprites don't seem to be transformed when drawing. I also should probably test the engine much more to ensure the internals are behaving as expected.