Begin adding Multi-Frame animated sprites.
Starting to add ability to create sprites that use sprite sheets to animate themselves. The way that this works is still up to change but is mostly similar to HaxeFlixel's FlxAnimationController.

Sprite and state have also begun to accept the type that should extend them so that other types can call the extended methods instead of the original. This isn't implemented yet but changes the feel of things.